chore(weave): address dataset_sources ch-wiring PR review feedback
Resolves the seven review threads from gtarpenning on PR #7241:

- Preserve explicitly-supplied empty-dict link_metadata (only None
  collapses to the empty-string sentinel).
- Derive call source_trace_id authoritatively from the call record and
  drop the client-supplied-trace_id requirement for call links.
- Use min(created_at) for the reverse-query first_seen_at so a blind
  relink cannot reset it.
- Translate added_by to the external user id in the forward-query response.
- Audit-log dataset_digest (an audit-only request field) on each link write.
- Extract all dataset_sources logic into dataset_sources/clickhouse.py
  (DatasetSourcesHandler); the trace server keeps thin delegators.
- Consolidate the redundant delete-edge-case integration tests.
