+++ weave/trace_server/external_to_internal_trace_server_adapter.py
@@ -892,9 +892,13 @@ class ExternalTraceServer(tsi.FullTraceServerInterface):
         req.project_id = self._idc.ext_to_int_project_id(req.project_id)
         if req.wb_user_id is not None:
             req.wb_user_id = self._idc.ext_to_int_user_id(req.wb_user_id)
-        return self._ref_apply(
+        res = self._ref_apply(
             self._internal_trace_server.dataset_sources_query, req, req.project_id
         )
+        for link in res.links:
+            if link.added_by is not None:
+                link.added_by = self._idc.int_to_ext_user_id(link.added_by)
+        return res
 
     def source_datasets_query(
         self, req: tsi.SourceDatasetsQueryReq
